device code to run on iot gateway hub
This code will run on a raspberry pie which represents the IOT edge gateway. This device connects, sends and receives instructions from the cloud. This device gathers json (and or bytes data) data from edge devices and using this software it can be converted to json and be sent through the cloud via wifi